[INLONG-4701] [Improve][Manager] Change the operating classes name from Operation to Operator
