+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/SinkOperatorFactory.java
@@ -15,34 +15,38 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.source;
+package org.apache.inlong.manager.service.sink;
 
 import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
-import org.apache.inlong.manager.common.enums.SourceType;
+import org.apache.inlong.manager.common.enums.SinkType;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
- * Factory for {@link StreamSourceOperation}.
+ * Factory for {@link StreamSinkOperator}.
  */
 @Service
-public class SourceOperationFactory {
+public class SinkOperatorFactory {
 
     @Autowired
-    private List<StreamSourceOperation> sourceOperationList;
+    private List<StreamSinkOperator> sinkOperatorList;
 
     /**
-     * Get a sink operation instance via the given sourceType
+     * Get a sink operator instance via the given sinkType
      */
-    public StreamSourceOperation getInstance(SourceType sourceType) {
-        return sourceOperationList.stream()
-                .filter(inst -> inst.accept(sourceType))
-                .findFirst()
-                .orElseThrow(() -> new BusinessException(ErrorCodeEnum.SOURCE_TYPE_NOT_SUPPORT,
-                        String.format(ErrorCodeEnum.SOURCE_TYPE_NOT_SUPPORT.getMessage(), sourceType)));
+    public StreamSinkOperator getInstance(SinkType sinkType) {
+        Optional<StreamSinkOperator> instance = sinkOperatorList.stream()
+                .filter(inst -> inst.accept(sinkType))
+                .findFirst();
+        if (!instance.isPresent()) {
+            throw new BusinessException(ErrorCodeEnum.SINK_TYPE_NOT_SUPPORT,
+                    String.format(ErrorCodeEnum.SINK_TYPE_NOT_SUPPORT.getMessage(), sinkType));
+        }
+        return instance.get();
     }
 
 }
