+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/hive/HiveSinkOperator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.sink.sqlserver;
+package org.apache.inlong.manager.service.sink.hive;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.pagehelper.Page;
@@ -28,10 +28,10 @@ import org.apache.inlong.manager.common.pojo.sink.SinkField;
 import org.apache.inlong.manager.common.pojo.sink.SinkListResponse;
 import org.apache.inlong.manager.common.pojo.sink.SinkRequest;
 import org.apache.inlong.manager.common.pojo.sink.StreamSink;
-import org.apache.inlong.manager.common.pojo.sink.sqlserver.SqlServerSink;
-import org.apache.inlong.manager.common.pojo.sink.sqlserver.SqlServerSinkDTO;
-import org.apache.inlong.manager.common.pojo.sink.sqlserver.SqlServerSinkListResponse;
-import org.apache.inlong.manager.common.pojo.sink.sqlserver.SqlServerSinkRequest;
+import org.apache.inlong.manager.common.pojo.sink.hive.HiveSink;
+import org.apache.inlong.manager.common.pojo.sink.hive.HiveSinkDTO;
+import org.apache.inlong.manager.common.pojo.sink.hive.HiveSinkListResponse;
+import org.apache.inlong.manager.common.pojo.sink.hive.HiveSinkRequest;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.common.util.Preconditions;
 import org.apache.inlong.manager.dao.entity.StreamSinkEntity;
@@ -48,12 +48,12 @@ import java.util.List;
 import java.util.function.Supplier;
 
 /**
- * SqlServer sink operation
+ * Hive sink operator
  */
 @Service
-public class SqlServerSinkOperation extends AbstractSinkOperator {
+public class HiveSinkOperator extends AbstractSinkOperator {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(SqlServerSinkOperation.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(HiveSinkOperator.class);
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -62,19 +62,22 @@ public class SqlServerSinkOperation extends AbstractSinkOperator {
 
     @Override
     public Boolean accept(SinkType sinkType) {
-        return SinkType.SQLSERVER.equals(sinkType);
+        return SinkType.HIVE == sinkType;
     }
 
     @Override
     public StreamSink getByEntity(@NotNull StreamSinkEntity entity) {
         Preconditions.checkNotNull(entity, ErrorCodeEnum.SINK_INFO_NOT_FOUND.getMessage());
         String existType = entity.getSinkType();
-        Preconditions.checkTrue(getSinkType().equals(existType),
-                String.format(ErrorCodeEnum.SINK_TYPE_NOT_SAME.getMessage(), getSinkType(), existType));
+        Preconditions.checkTrue(this.getSinkType().equals(existType),
+                String.format(ErrorCodeEnum.SINK_TYPE_NOT_SAME.getMessage(), this.getSinkType(), existType));
+
         StreamSink response = this.getFromEntity(entity, this::getSink);
         List<StreamSinkFieldEntity> entities = sinkFieldMapper.selectBySinkId(entity.getId());
-        List<SinkField> infos = CommonBeanUtils.copyListProperties(entities, SinkField::new);
+        List<SinkField> infos = CommonBeanUtils.copyListProperties(entities,
+                SinkField::new);
         response.setSinkFieldList(infos);
+
         return response;
     }
 
@@ -84,11 +87,12 @@ public class SqlServerSinkOperation extends AbstractSinkOperator {
         if (entity == null) {
             return result;
         }
-        String existType = entity.getSinkType();
-        Preconditions.checkTrue(getSinkType().equals(existType),
-                String.format(ErrorCodeEnum.SINK_TYPE_NOT_SAME.getMessage(), getSinkType(), existType));
 
-        SqlServerSinkDTO dto = SqlServerSinkDTO.getFromJson(entity.getExtParams());
+        String existType = entity.getSinkType();
+        Preconditions.checkTrue(this.getSinkType().equals(existType),
+                String.format(ErrorCodeEnum.SINK_TYPE_NOT_SAME.getMessage(), this.getSinkType(), existType));
+
+        HiveSinkDTO dto = HiveSinkDTO.getFromJson(entity.getExtParams());
         CommonBeanUtils.copyProperties(entity, result, true);
         CommonBeanUtils.copyProperties(dto, result, true);
 
@@ -100,16 +104,16 @@ public class SqlServerSinkOperation extends AbstractSinkOperator {
         if (CollectionUtils.isEmpty(entityPage)) {
             return new PageInfo<>();
         }
-        return entityPage.toPageInfo(entity -> this.getFromEntity(entity, SqlServerSinkListResponse::new));
+        return entityPage.toPageInfo(entity -> this.getFromEntity(entity, HiveSinkListResponse::new));
     }
 
     @Override
     protected void setTargetEntity(SinkRequest request, StreamSinkEntity targetEntity) {
-        Preconditions.checkTrue(getSinkType().equals(request.getSinkType()),
+        Preconditions.checkTrue(this.getSinkType().equals(request.getSinkType()),
                 ErrorCodeEnum.SINK_TYPE_NOT_SUPPORT.getMessage() + ": " + getSinkType());
-        SqlServerSinkRequest sinkRequest = (SqlServerSinkRequest) request;
+        HiveSinkRequest sinkRequest = (HiveSinkRequest) request;
         try {
-            SqlServerSinkDTO dto = SqlServerSinkDTO.getFromRequest(sinkRequest);
+            HiveSinkDTO dto = HiveSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
             LOGGER.error("parsing json string to sink info failed", e);
@@ -119,11 +123,12 @@ public class SqlServerSinkOperation extends AbstractSinkOperator {
 
     @Override
     protected String getSinkType() {
-        return SinkType.SINK_SQLSERVER;
+        return SinkType.SINK_HIVE;
     }
 
     @Override
     protected StreamSink getSink() {
-        return new SqlServerSink();
+        return new HiveSink();
     }
+
 }
