+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/source/sqlserver/SqlServerSourceOperator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.source.autopush;
+package org.apache.inlong.manager.service.source.sqlserver;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.pagehelper.Page;
@@ -27,61 +27,61 @@ import org.apache.inlong.manager.common.exceptions.BusinessException;
 import org.apache.inlong.manager.common.pojo.source.SourceListResponse;
 import org.apache.inlong.manager.common.pojo.source.SourceRequest;
 import org.apache.inlong.manager.common.pojo.source.StreamSource;
-import org.apache.inlong.manager.common.pojo.source.autopush.AutoPushSource;
-import org.apache.inlong.manager.common.pojo.source.autopush.AutoPushSourceDTO;
-import org.apache.inlong.manager.common.pojo.source.autopush.AutoPushSourceListResponse;
-import org.apache.inlong.manager.common.pojo.source.autopush.AutoPushSourceRequest;
+import org.apache.inlong.manager.common.pojo.source.sqlserver.SqlServerSource;
+import org.apache.inlong.manager.common.pojo.source.sqlserver.SqlServerSourceDTO;
+import org.apache.inlong.manager.common.pojo.source.sqlserver.SqlServerSourceListResponse;
+import org.apache.inlong.manager.common.pojo.source.sqlserver.SqlServerSourceRequest;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.common.util.Preconditions;
 import org.apache.inlong.manager.dao.entity.StreamSourceEntity;
-import org.apache.inlong.manager.service.source.AbstractSourceOperation;
+import org.apache.inlong.manager.service.source.AbstractSourceOperator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.function.Supplier;
 
 /**
- * DataProxy SDK source operation
+ * SqlServer source operator
  */
 @Service
-public class AutoPushSourceOperation extends AbstractSourceOperation {
+public class SqlServerSourceOperator extends AbstractSourceOperator {
 
     @Autowired
     private ObjectMapper objectMapper;
+
+    @Override
+    public Boolean accept(SourceType sourceType) {
+        return SourceType.SQLSERVER == sourceType;
+    }
+
+    @Override
+    protected String getSourceType() {
+        return SourceType.SQLSERVER.getType();
+    }
+
+    @Override
+    protected StreamSource getSource() {
+        return new SqlServerSource();
+    }
 
     @Override
     public PageInfo<? extends SourceListResponse> getPageInfo(Page<StreamSourceEntity> entityPage) {
         if (CollectionUtils.isEmpty(entityPage)) {
             return new PageInfo<>();
         }
-        return entityPage.toPageInfo(entity -> this.getFromEntity(entity, AutoPushSourceListResponse::new));
+        return entityPage.toPageInfo(entity -> this.getFromEntity(entity, SqlServerSourceListResponse::new));
     }
 
     @Override
     protected void setTargetEntity(SourceRequest request, StreamSourceEntity targetEntity) {
-        AutoPushSourceRequest sourceRequest = (AutoPushSourceRequest) request;
+        SqlServerSourceRequest sourceRequest = (SqlServerSourceRequest) request;
         CommonBeanUtils.copyProperties(sourceRequest, targetEntity, true);
         try {
-            AutoPushSourceDTO dto = AutoPushSourceDTO.getFromRequest(sourceRequest);
+            SqlServerSourceDTO dto = SqlServerSourceDTO.getFromRequest(sourceRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
             throw new BusinessException(ErrorCodeEnum.SOURCE_INFO_INCORRECT.getMessage());
         }
-    }
-
-    @Override
-    protected String getSourceType() {
-        return SourceType.AUTO_PUSH.getType();
-    }
-
-    @Override
-    protected StreamSource getSource() {
-        return new AutoPushSource();
-    }
-
-    @Override
-    public Boolean accept(SourceType sourceType) {
-        return SourceType.AUTO_PUSH == sourceType;
     }
 
     @Override
@@ -93,7 +93,7 @@ public class AutoPushSourceOperation extends AbstractSourceOperation {
         String existType = entity.getSourceType();
         Preconditions.checkTrue(getSourceType().equals(existType),
                 String.format(ErrorCodeEnum.SOURCE_TYPE_NOT_SAME.getMessage(), getSourceType(), existType));
-        AutoPushSourceDTO dto = AutoPushSourceDTO.getFromJson(entity.getExtParams());
+        SqlServerSourceDTO dto = SqlServerSourceDTO.getFromJson(entity.getExtParams());
         CommonBeanUtils.copyProperties(entity, result, true);
         CommonBeanUtils.copyProperties(dto, result, true);
         return result;
